refactorings for single model module, but with errors of
dataclasses_json.undefined.UndefinedParameterError: No field of type dataclasses_json.CatchAll defined
