+++ tests/integration/test_api_message.py
@@ -1,7 +1,5 @@
 import datetime as _dt
 
-import pytest as _pt
-
 import mailpit.client.api as _api
